LPS-129316 Rename banners to banner
