+++ modules/apps/fragment/fragment-collection-contributor/fragment-collection-contributor-cookie-banner/bnd.bnd
@@ -1,0 +1,5 @@
+Bundle-Name: Liferay Fragment Collection Contributor Cookie Banner
+Bundle-SymbolicName: com.liferay.fragment.collection.contributor.cookie.banner
+Bundle-Version: 1.0.0
+Web-ContextPath: /fragment-collection-contributor-cookie-banner
+-dsannotations-options: inherit
