+++ modules/apps/fragment/fragment-collection-contributor/fragment-collection-contributor-cookie-banner/build.gradle
