+++ modules/apps/fragment/fragment-collection-contributor/fragment-collection-contributor-cookie-banner/src/main/resources/META-INF/resources/thumbnails/cookie_banner_simple.png
