+++ modules/apps/fragment/fragment-collection-contributor/fragment-collection-contributor-cookie-banner/src/main/resources/com/liferay/fragment/collection/contributor/cookie/banner/dependencies/collection.json
@@ -1,0 +1,3 @@
+{
+	"name": "cookie-banner"
+}
