+++ modules/apps/fragment/fragment-collection-contributor/fragment-collection-contributor-cookie-banner/src/main/resources/com/liferay/fragment/collection/contributor/cookie/banner/dependencies/cookie-banner-simple/fragment.json
