+++ modules/apps/fragment/fragment-collection-contributor/fragment-collection-contributor-cookie-banner/src/main/resources/content/Language.properties
@@ -1,0 +1,2 @@
+cookie-banner=Cookie Banner
+cookie-banner-simple=Cookie Banner Simple
