--- modules/apps/fragment/fragment-collection-contributor/fragment-collection-contributor-cookie-banners/bnd.bnd
@@ -1,5 +1,0 @@
-Bundle-Name: Liferay Fragment Collection Contributor Cookie Banners
-Bundle-SymbolicName: com.liferay.fragment.collection.contributor.cookie.banners
-Bundle-Version: 1.0.0
-Web-ContextPath: /fragment-collection-contributor-cookie-banners
--dsannotations-options: inherit
