--- modules/apps/fragment/fragment-collection-contributor/fragment-collection-contributor-cookie-banners/src/main/resources/com/liferay/fragment/collection/contributor/cookie/banners/dependencies/collection.json
@@ -1,3 +1,0 @@
-{
-	"name": "cookie-banners"
-}
