--- modules/apps/fragment/fragment-collection-contributor/fragment-collection-contributor-cookie-banners/src/main/resources/content/Language.properties
@@ -1,2 +1,0 @@
-cookie-banner-simple=Cookie Banner Simple
-cookie-banners=Cookie Banners
